+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 -   Adds `IconNames` enum to make `allSVGs` array globally accessible
 -   Adds `width: 100%` globally to SVGs
 -   `Card` now constrains items in its `image` column to the column width
+-   Adds the `forwardRef` implementation to the `Input` component
 
 ### Changes